+++ IS4.AuthorizationCenter/Controllers/Account/AccountController.cs
@@ -9,7 +9,7 @@ using IdentityServer4.Extensions;
 using IdentityServer4.Models;
 using IdentityServer4.Services;
 using IdentityServer4.Stores;
-using IdentityServer4.Test;
+using IS4.AuthorizationCenter.Controllers.Authorizations;
 using IS4.AuthorizationCenter.Models.Account;
 using IS4.AuthorizationCenter.Models.Entity;
 using IS4.AuthorizationCenter.Models.Options;
@@ -28,7 +28,7 @@ using System.Threading.Tasks;
 namespace IS4.AuthorizationCenter.Controllers.Account
 {
     /// <summary>
-    ///这个控制器为本地和外部帐户实现了一个典型的登录/注销/注册/角色/用户管理。
+    ///这个控制器为本地和外部帐户实现了一个典型的登录/注销/注册。
     ///登录服务封装了与用户数据存储的交互。
     ///交互服务为UI与identityserver进行验证和上下文检索提供了一种通信方式
     [SecurityHeaders]
@@ -62,91 +62,6 @@ namespace IS4.AuthorizationCenter.Controllers.Account
             _events = events;
             _accountOptions = accountOptions.Value;
             _roleManager = roleManager;
-        }
-
-        /// <summary>
-        /// 显示登陆页面
-        /// </summary>
-        /// <param name="returnUrl">返回的url</param>
-        /// <returns></returns>
-        [HttpGet]
-        public async Task<IActionResult> Login(string returnUrl)
-        {
-            // 创建登陆页面使用的数据模型
-            LoginViewModel vm = await BuildLoginViewModelAsync(returnUrl);
-
-            //如果是外部登陆
-            if (vm.IsExternalLoginOnly)
-            {
-                // 跳转到外部登陆页面
-                return RedirectToAction("Challenge", "External", new { provider = vm.ExternalLoginScheme, returnUrl });
-            }
-
-            return View(vm);
-        }
-
-        /// <summary>
-        /// 处理用户名/密码登录的回发
-        /// </summary>
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Login(LoginInputModel model, string button)
-        {
-            if (button != "login")
-            {
-                // 用户点击取消按钮
-                var context = await _interaction.GetAuthorizationContextAsync(model.ReturnUrl);
-                if (context != null)
-                {
-                    //如果用户取消，将结果发送回IdentityServer
-                    //拒绝同意(即使这个客户不需要同意)。
-                    //这将向客户端发送一个拒绝访问的OIDC错误响应。
-                    await _interaction.GrantConsentAsync(context, ConsentResponse.Denied);
-
-                    //我们可以信任模型。ReturnUrl因为GetAuthorizationContextAsync返回非空
-                    return Redirect(model.ReturnUrl);
-                }
-                else
-                {
-                    // 因为我们没有有效的上下文，所以我们只能返回到主页
-                    return Redirect("~/");
-                }
-            }
-
-            if (ModelState.IsValid)
-            {
-                //通过用户名查找用户
-                var user = await _userManager.FindByNameAsync(model.Username);
-
-
-                if (!user.IsDelete)
-                {
-                    //通过密码登录
-                    var result = await _signInManager.PasswordSignInAsync(model.Username, model.Password, model.RememberLogin, lockoutOnFailure: true);
-                    if (result.Succeeded)
-                    {
-                        //创建登录成功事件
-                        await _events.RaiseAsync(new UserLoginSuccessEvent(user.UserName, user.Id.ToString(), user.UserName));
-
-                        //确保returnUrl仍然有效，如果是，则重定向回authorize endpoint或本地页面
-                        // IsLocalUrl检查是必要的，如果你想支持额外的本地页面，否则IsValidReturnUrl是更严格的
-                        if (_interaction.IsValidReturnUrl(model.ReturnUrl) || Url.IsLocalUrl(model.ReturnUrl))
-                        {
-                            return Redirect(model.ReturnUrl);
-                        }
-
-                        return Redirect("~/");
-                    }
-                }
-
-                await _events.RaiseAsync(new UserLoginFailureEvent(model.Username, _accountOptions.InvalidUserErrorMessage));
-
-                ModelState.AddModelError("", _accountOptions.InvalidCredentialsErrorMessage);
-            }
-
-            // 出了差错，用错误来表示
-            var vm = await BuildLoginViewModelAsync(model);
-            return View(vm);
         }
 
         /// <summary>
@@ -479,207 +394,6 @@ namespace IS4.AuthorizationCenter.Controllers.Account
             return View();
         }
 
-        #region 角色管理
-
-        /// <summary>
-        /// 添加角色
-        /// </summary>
-        /// <param name="returnUrl"></param>
-        /// <returns></returns>
-        [HttpGet]
-        [Route("account/Roleregister")]
-        public IActionResult RoleRegister(string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            return View();
-        }
-
-        /// <summary>
-        /// 添加角色
-        /// </summary>
-        /// <param name="model"></param>
-        /// <param name="returnUrl"></param>
-        /// <returns></returns>
-        [HttpPost]
-        [Route("account/Roleregister")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> RoleRegister(RoleRegisterViewModel model, string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            IdentityResult result = new IdentityResult();
-
-            if (ModelState.IsValid)
-            {
-                //通过角色名查找角色
-                var roleItem = await _roleManager.FindByNameAsync(model.RoleName);
-
-                if (roleItem == null)
-                {
-
-                    var role = new ApplicationRole
-                    {
-                        Name = model.RoleName
-                    };
-
-
-                    result = await _roleManager.CreateAsync(role);
-
-                    if (result.Succeeded)
-                    {
-
-                        if (result.Succeeded)
-                        {
-                            return RedirectToLocal(returnUrl);
-                        }
-                    }
-
-                }
-                else
-                {
-                    ModelState.AddModelError(string.Empty, $"{roleItem?.Name} 已经存在");
-
-                }
-
-                AddErrors(result);
-            }
-
-            // 如果我们走到这一步，有什么失败了，重新显示
-            return View(model);
-        }
-
-        /// <summary>
-        /// 角色列表
-        /// </summary>
-        /// <param name="returnUrl"></param>
-        /// <returns></returns>
-        [HttpGet]
-        [Route("account/Roles")]
-        [Authorize]
-        public IActionResult Roles(string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            var roles = _roleManager.Roles.Where(d => !d.IsDeleted).ToList();
-
-            return View(roles);
-        }
-
-        /// <summary>
-        /// 编辑角色
-        /// </summary>
-        /// <param name="id"></param>
-        /// <param name="returnUrl"></param>
-        /// <returns></returns>
-        [HttpGet("{id}")]
-        [Route("account/Roleedit/{id}")]
-        [Authorize(Roles = "SuperAdmin")]
-        public async Task<IActionResult> RoleEdit(string id, string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _roleManager.FindByIdAsync(id);
-
-            if (user == null)
-            {
-                return NotFound();
-            }
-
-            return View(new RoleEditViewModel(user.Id.ToString(), user.Name));
-        }
-
-        /// <summary>
-        /// 编辑角色
-        /// </summary>
-        /// <param name="model"></param>
-        /// <param name="id"></param>
-        /// <param name="returnUrl"></param>
-        /// <returns></returns>
-        [HttpPost]
-        [Route("account/Roleedit/{id}")]
-        [ValidateAntiForgeryToken]
-        [Authorize(Roles = "SuperAdmin")]
-        public async Task<IActionResult> RoleEdit(RoleEditViewModel model, string id, string returnUrl = null)
-        {
-            ViewData["ReturnUrl"] = returnUrl;
-            IdentityResult result = new IdentityResult();
-
-            if (ModelState.IsValid)
-            {
-                //通过角色id查询角色
-                var roleItem =await _roleManager.FindByIdAsync(model.Id);
-
-                if (roleItem != null)
-                {
-                    roleItem.Name = model.RoleName;
-
-                    result = await _roleManager.UpdateAsync(roleItem);
-
-                    if (result.Succeeded)
-                    {
-                        return RedirectToLocal(returnUrl);
-                    }
-
-                }
-                else
-                {
-                    ModelState.AddModelError(string.Empty, $"{roleItem?.Name} no exist!");
-                }
-
-                AddErrors(result);
-            }
-
-            // 如果我们走到这一步，有什么失败了，重新显示
-            return View(model);
-        }
-
-
-        /// <summary>
-        /// 删除角色
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        [HttpPost]
-        [Route("account/Roledelete/{id}")]
-        [Authorize(Roles = "SuperAdmin")]
-        public async Task<JsonResult> RoleDelete(string id)
-        {
-            IdentityResult result = new IdentityResult();
-
-            if (ModelState.IsValid)
-            {
-                var roleItem = _roleManager.FindByIdAsync(id).Result;
-
-                if (roleItem != null)
-                {
-                    roleItem.IsDeleted = true;
-
-
-                    result = await _roleManager.UpdateAsync(roleItem);
-
-                    if (result.Succeeded)
-                    {
-                        return Json(result);
-                    }
-
-                }
-                else
-                {
-                    ModelState.AddModelError(string.Empty, $"{roleItem?.Name} 不存在");
-                }
-
-                AddErrors(result);
-            }
-
-            return Json(result.Errors);
-
-        }
-
-        #endregion
-
-
         /// <summary>
         /// 显示注销页
         /// </summary>
@@ -744,78 +458,9 @@ namespace IS4.AuthorizationCenter.Controllers.Account
         /* helper APIs for the AccountController */
         /*****************************************/
 
-        /// <summary>
-        /// 建立登录展示模型
-        /// </summary>
-        /// <param name="returnUrl"></param>
-        /// <returns></returns>
-        private async Task<LoginViewModel> BuildLoginViewModelAsync(string returnUrl)
-        {
-            var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
-            if (context?.IdP != null && await _schemeProvider.GetSchemeAsync(context.IdP) != null)
-            {
-                var local = context.IdP == IdentityServerConstants.LocalIdentityProvider;
+        
 
-                // 这意味着要使UI短路，只触发一个外部IdP
-                var vm = new LoginViewModel
-                {
-                    EnableLocalLogin = local,
-                    ReturnUrl = returnUrl,
-                    Username = context?.LoginHint,
-                };
-
-                if (!local)
-                {
-                    vm.ExternalProviders = new[] { new ExternalProvider { AuthenticationScheme = context.IdP } };
-                }
-
-                return vm;
-            }
-
-            var schemes = await _schemeProvider.GetAllSchemesAsync();
-
-            var providers = schemes
-                .Where(x => x.DisplayName != null ||
-                            x.Name.Equals(_accountOptions.WindowsAuthenticationSchemeName, StringComparison.OrdinalIgnoreCase)
-                )
-                .Select(x => new ExternalProvider
-                {
-                    DisplayName = x.DisplayName ?? x.Name,
-                    AuthenticationScheme = x.Name
-                }).ToList();
-
-            var allowLocal = true;
-            if (context?.ClientId != null)
-            {
-                var client = await _clientStore.FindEnabledClientByIdAsync(context.ClientId);
-                if (client != null)
-                {
-                    allowLocal = client.EnableLocalLogin;
-
-                    if (client.IdentityProviderRestrictions != null && client.IdentityProviderRestrictions.Any())
-                    {
-                        providers = providers.Where(provider => client.IdentityProviderRestrictions.Contains(provider.AuthenticationScheme)).ToList();
-                    }
-                }
-            }
-
-            return new LoginViewModel
-            {
-                AllowRememberLogin = _accountOptions.AllowRememberLogin,
-                EnableLocalLogin = allowLocal && _accountOptions.AllowLocalLogin,
-                ReturnUrl = returnUrl,
-                Username = context?.LoginHint,
-                ExternalProviders = providers.ToArray()
-            };
-        }
-
-        private async Task<LoginViewModel> BuildLoginViewModelAsync(LoginInputModel model)
-        {
-            var vm = await BuildLoginViewModelAsync(model.ReturnUrl);
-            vm.Username = model.Username;
-            vm.RememberLogin = model.RememberLogin;
-            return vm;
-        }
+      
 
         /// <summary>
         /// 建立登出展示模型
@@ -899,7 +544,7 @@ namespace IS4.AuthorizationCenter.Controllers.Account
             }
             else
             {
-                return RedirectToAction(nameof(AccountController.Login), "Account");
+                return RedirectToAction(nameof(OAuth2Controller.Authorization), "OAuth2");
             }
         }
